pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paper-digest"
version = "0.1.0"
description = "基于 Zotero + Qwen 的科研文献知识编译系统"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Author"}
]
dependencies = [
    "pyzotero>=1.5.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
]

[project.scripts]
paper-digest = "paper_digest.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["paper_digest"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.uv.workspace]
members = [
    ".",
]

[tool.uv.sources]
paper-digest = { workspace = true }

[dependency-groups]
dev = [
    "paper-digest",
]
